Exceptions customizadas

+++ src/main/java/tech/curtiu/brcommerce/services/ProductService.java
@@ -4,12 +4,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import jakarta.persistence.EntityNotFoundException;
 import tech.curtiu.brcommerce.dto.ProductDTO;
 import tech.curtiu.brcommerce.entities.Product;
 import tech.curtiu.brcommerce.mappers.ProductMapper;
 import tech.curtiu.brcommerce.repositories.ProductRepository;
+import tech.curtiu.brcommerce.services.exceptions.DataIntegrityViolationException;
 import tech.curtiu.brcommerce.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -43,15 +46,26 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product product = productRepository.getReferenceById(id);
-        productMapper.updateEntityFromDto(dto, product);
-        product = productRepository.save(product);
-        return productMapper.toDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            productMapper.updateEntityFromDto(dto, product);
+            product = productRepository.save(product);
+            return productMapper.toDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Product not found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Product not found");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (Exception e) {
+            throw new DataIntegrityViolationException("Integrity violation");
+        }
     }
 
 }
